+++ app/agent/__init__.py
@@ -6,6 +6,7 @@ from ..agent.planning import PlanningAgent
 from ..agent.react import ReActAgent
 from ..agent.swe import SWEAgent
 from ..agent.toolcall import ToolCallAgent
+from ..agent.manus import ManusAgent
 
 
 __all__ = [
@@ -17,4 +18,5 @@ __all__ = [
     "SWEAgent",
     "ToolCallAgent",
     "MCPAgent",
+    "ManusAgent",
 ]
