+++ app/tool/__init__.py
@@ -1,15 +1,18 @@
-from .base import BaseTool
-from .bash import Bash
-from .browser_use_tool import BrowserUseTool
-from .create_chat_completion import CreateChatCompletion
-from .planning import PlanningTool
-from .str_replace_editor import StrReplaceEditor
-from .terminate import Terminate
-from .tool_collection import ToolCollection
-
+from ..tool.base import BaseTool, ToolResult, CLIResult, ToolFailure
+from ..tool.bash import Bash
+from ..tool.browser_use_tool import BrowserUseTool
+from ..tool.create_chat_completion import CreateChatCompletion
+from ..tool.planning import PlanningTool
+from ..tool.str_replace_editor import StrReplaceEditor
+from ..tool.terminate import Terminate
+from ..tool.tool_collection import ToolCollection
+from ..tool.web_search import WebSearch
+from ..tool.file_operators import FileOperator, LocalFileOperator, PathLike, SandboxFileOperator
+from ..tool.file_saver import FileSaver
 
 __all__ = [
     "BaseTool",
+    "ToolResult",
     "Bash",
     "BrowserUseTool",
     "Terminate",
@@ -17,4 +20,12 @@ __all__ = [
     "ToolCollection",
     "CreateChatCompletion",
     "PlanningTool",
+    "WebSearch",
+    "CLIResult",
+    "FileOperator",
+    "LocalFileOperator",
+    "PathLike",
+    "SandboxFileOperator",
+    "FileSaver",
+    "ToolFailure",
 ]
