+++ app/tool/search/__init__.py
@@ -1,5 +1,5 @@
 from .baidu_search import BaiduSearchEngine
-from ..base import WebSearchEngine
+from .base import WebSearchEngine
 from .bing_search import BingSearchEngine
 from .duckduckgo_search import DuckDuckGoSearchEngine
 from .google_search import GoogleSearchEngine
