+++ app/tool/str_replace_editor.py
@@ -6,13 +6,9 @@ from typing import Any, DefaultDict, List, Literal, Optional, get_args
 
 from ..config import config
 from ..exceptions import ToolError
-from .base import CLIResult, ToolResult, BaseTool
-from .file_operators import (
-    FileOperator,
-    LocalFileOperator,
-    PathLike,
-    SandboxFileOperator,
-)
+from ..tool import CLIResult, ToolResult, BaseTool
+from ..tool import FileOperator, LocalFileOperator, PathLike, SandboxFileOperator
+
 
 
 Command = Literal[
