+++ app/tool/tool_collection.py
@@ -2,7 +2,7 @@
 from typing import Any, Dict, List
 
 from ..exceptions import ToolError
-from .base import BaseTool, ToolFailure, ToolResult
+from ..tool import BaseTool, ToolFailure, ToolResult
 
 
 class ToolCollection:
