+++ app/tool/web_search.py
@@ -5,8 +5,8 @@ from tenacity import retry, stop_after_attempt, wait_exponential
 
 from ..config import config
 from ..logger import logger
-from .base import BaseTool
-from .search import (
+from ..tool.base import BaseTool
+from ..tool.search import (
     BaiduSearchEngine,
     BingSearchEngine,
     DuckDuckGoSearchEngine,
